requirements.txt
requests==2.28.1
hikari==2.0.0.dev115
python-dotenv==0.21.0
finnhub-python==2.4.14
polygon-api-client===1.12.4